Create abstractHooks
