+++ packages/core/src/index.ts
@@ -156,3 +156,4 @@ export { ErrorHandling, ErrorSource } from './domain/error/error.types'
 export * from './domain/deflate'
 export * from './domain/connectivity'
 export * from './tools/stackTrace/handlingStack'
+export * from './tools/abstractHooks'
