+++ packages/rum-core/src/boot/startRum.spec.ts
@@ -44,9 +44,9 @@ import type { RumConfiguration } from '../domain/configuration'
 import { RumEventType } from '../rawRumEvent.types'
 import type { PageStateHistory } from '../domain/contexts/pageStateHistory'
 import { createCustomVitalsState } from '../domain/vital/vitalCollection'
-import { createHooks } from '../hooks'
 import { startUrlContexts } from '../domain/contexts/urlContexts'
 import { startSessionContext } from '../domain/contexts/sessionContext'
+import { createHooks } from '../domain/hooks'
 import { startRum, startRumEventCollection } from './startRum'
 
 function collectServerEvents(lifeCycle: LifeCycle) {
