+++ packages/rum-core/src/domain/contexts/accountContext.spec.ts
@@ -1,9 +1,9 @@
 import type { ContextManager, RelativeTime } from '@datadog/browser-core'
-import { display, removeStorageListeners } from '@datadog/browser-core'
+import { display, HookNames, removeStorageListeners } from '@datadog/browser-core'
 import { registerCleanupTask } from '@datadog/browser-core/test'
 import { mockRumConfiguration } from '../../../test'
-import type { Hooks } from '../../hooks'
-import { HookNames, createHooks } from '../../hooks'
+import type { Hooks } from '../hooks'
+import { createHooks } from '../hooks'
 import { startAccountContext } from './accountContext'
 
 describe('account context', () => {
