+++ packages/rum-core/src/domain/contexts/ciVisibilityContext.ts
@@ -1,8 +1,7 @@
-import { getInitCookie } from '@datadog/browser-core'
+import { getInitCookie, HookNames, SKIPPED } from '@datadog/browser-core'
 import { createCookieObservable } from '../../browser/cookieObservable'
-import type { Hooks, DefaultRumEventAttributes } from '../../hooks'
-import { SKIPPED, HookNames } from '../../hooks'
 import { SessionType } from '../rumSessionManager'
+import type { DefaultRumEventAttributes, Hooks } from '../hooks'
 
 export const CI_VISIBILITY_TEST_ID_COOKIE_NAME = 'datadog-ci-visibility-test-execution-id'
