+++ packages/rum-core/src/domain/contexts/connectivityContext.spec.ts
@@ -1,12 +1,15 @@
 import { setNavigatorOnLine, setNavigatorConnection } from '@datadog/browser-core/test'
+import { HookNames } from '@datadog/browser-core'
 import type { RelativeTime } from '@datadog/browser-core'
-import type { Hooks } from '../../hooks'
-import { createHooks, HookNames } from '../../hooks'
+import type { Hooks } from '../hooks'
+import { createHooks } from '../hooks'
 import { startConnectivityContext } from './connectivityContext'
 
 describe('startConnectivityContext', () => {
   describe('assemble hook', () => {
-    let hooks: Hooks
+    type NewType = Hooks
+
+    let hooks: NewType
 
     beforeEach(() => {
       hooks = createHooks()
