+++ packages/rum-core/src/domain/contexts/defaultContext.ts
@@ -1,7 +1,6 @@
-import { canUseEventBridge, currentDrift, round, timeStampNow } from '@datadog/browser-core'
-import type { Hooks, DefaultRumEventAttributes } from '../../hooks'
-import { HookNames } from '../../hooks'
+import { canUseEventBridge, currentDrift, HookNames, round, timeStampNow } from '@datadog/browser-core'
 import type { RumConfiguration } from '../configuration'
+import type { DefaultRumEventAttributes, Hooks } from '../hooks'
 
 // replaced at build time
 declare const __BUILD_ENV__SDK_VERSION__: string
