+++ packages/rum-core/src/domain/contexts/featureFlagContext.spec.ts
@@ -1,13 +1,13 @@
 import type { RelativeTime } from '@datadog/browser-core'
-import { relativeToClocks } from '@datadog/browser-core'
+import { HookNames, relativeToClocks } from '@datadog/browser-core'
 import type { Clock } from '@datadog/browser-core/test'
 import { mockClock, registerCleanupTask } from '@datadog/browser-core/test'
 import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
 import type { ViewCreatedEvent, ViewEndedEvent } from '../view/trackViews'
-import type { Hooks } from '../../hooks'
-import { createHooks, HookNames } from '../../hooks'
 import type { RumConfiguration } from '../configuration'
 import { RumEventType } from '../../rawRumEvent.types'
+import type { Hooks } from '../hooks'
+import { createHooks } from '../hooks'
 import type { FeatureFlagContexts } from './featureFlagContext'
 import { startFeatureFlagContexts } from './featureFlagContext'
