+++ packages/rum-core/src/domain/contexts/pageStateHistory.spec.ts
@@ -1,11 +1,12 @@
 import type { ServerDuration, Duration, RelativeTime } from '@datadog/browser-core'
+import { HookNames } from '@datadog/browser-core'
 import type { Clock } from '../../../../core/test'
 import { mockClock, registerCleanupTask } from '../../../../core/test'
 import { mockRumConfiguration, mockPerformanceObserver } from '../../../test'
-import { createHooks, HookNames } from '../../hooks'
-import type { Hooks } from '../../hooks'
 import { RumEventType } from '../../rawRumEvent.types'
 import * as performanceObservable from '../../browser/performanceObservable'
+import type { Hooks } from '../hooks'
+import { createHooks } from '../hooks'
 import type { PageStateHistory } from './pageStateHistory'
 import { PageState, startPageStateHistory } from './pageStateHistory'
