+++ packages/rum-core/src/domain/view/viewCollection.spec.ts
@@ -1,4 +1,4 @@
-import { Observable } from '@datadog/browser-core'
+import { DISCARDED, HookNames, Observable } from '@datadog/browser-core'
 import type { Duration, RelativeTime, ServerDuration, TimeStamp } from '@datadog/browser-core'
 import { mockClock, registerCleanupTask } from '@datadog/browser-core/test'
 import type { RecorderApi } from '../../boot/rumPublicApi'
@@ -9,9 +9,9 @@ import type { RawRumEventCollectedData } from '../lifeCycle'
 import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
 import type { RumConfiguration } from '../configuration'
 import type { LocationChange } from '../../browser/locationChangeObservable'
-import type { Hooks } from '../../hooks'
-import { DISCARDED, HookNames, createHooks } from '../../hooks'
 import type { ViewHistoryEntry } from '../contexts/viewHistory'
+import type { Hooks } from '../hooks'
+import { createHooks } from '../hooks'
 import { startViewCollection } from './viewCollection'
 import type { ViewEvent } from './trackViews'
